Fix branch_name in CI and refactor Env out
